+++ BALSAMIC/constants/cache.py
@@ -390,7 +390,7 @@ REFERENCE_FILES: Dict[GenomeVersion, Dict[str, dict]] = {
             "url": "https://ftp.ncbi.nlm.nih.gov/pub/clinvar/vcf_GRCh38/clinvar.vcf.gz",
             "file_type": FileType.VCF,
             "gzip": True,
-            "file_name": "clinvar.vcf",
+            "file_name": "clinvar_preprocess.vcf",
             "dir_name": "variants",
         },
         "somalier_sites": {