+++ CHANGELOG.rst
@@ -8,6 +8,8 @@ Added:
 * logfile for balsamic wrapper https://github.com/Clinical-Genomics/BALSAMIC/pull/1558
 * analysis status text file for easier prodbioinfo handling https://github.com/Clinical-Genomics/BALSAMIC/pull/1558
 * requested memory to each rule https://github.com/Clinical-Genomics/BALSAMIC/pull/1558
+* added rule and script for adding CLNVID to clinvar during cache installation https://github.com/Clinical-Genomics/BALSAMIC/pull/1576
+
 
 Changed:
 ^^^^^^^^